# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel - local'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/v1/user-detail:
    get:
      summary: Profile
      operationId: profile
      description: 'Profile Detail'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses: {  }
      tags:
        - Auth
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []